+++ packages/assets/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@synergy-design-system/assets-v1.3.0](https://github.com/synergy-design-system/synergy-design-system/compare/assets/1.2.1...assets/1.3.0) (2024-01-30)
+
+
+### Features
+
+* ✨ syn-select / syn-option / syn-optgroup ([#274](https://github.com/synergy-design-system/synergy-design-system/issues/274)) ([25c6788](https://github.com/synergy-design-system/synergy-design-system/commit/25c678829e58a173c0fc23005a4f724b6d792dd7))
+
 # [@synergy-design-system/assets-v1.2.1](https://github.com/synergy-design-system/synergy-design-system/compare/assets/1.2.0...assets/1.2.1) (2024-01-24)